Move qvm-prefs parser to toolparsers
